+++ src/main/kotlin/com/phpusr/wildrace/dto/RunnerDto.kt
@@ -7,6 +7,6 @@ import com.phpusr.wildrace.domain.Views
 @JsonView(Views.StatDtoREST::class)
 class RunnerDto(
         val profile: Profile,
-        val numberOfRuns: Long,
-        val sumDistance: Long
+        val numberOfRuns: Int,
+        val sumDistance: Int
 )
